// packages/core/src/types/_aux.d.ts
export interface Aux<Stage extends ModelStage, Element extends string, Deployment extends string, View extends string, Project extends string, Spec extends AnySpec> {
    Stage: Stage;
    ProjectId: Project;
    ElementId: Element;
    DeploymentId: Deployment;
    ViewId: View;
    ElementKind: Spec['ElementKind'];
    DeploymentKind: Spec['DeploymentKind'];
    RelationKind: Spec['RelationKind'];
    Tag: Spec['Tag'];
    MetadataKey: Spec['MetadataKey'];
}
export type AnyOnStage<Stage extends ModelStage> = Aux<Stage, any, any, any, any, any>;
export type AnyParsed = AnyOnStage<'parsed'>;
export type AnyComputed = AnyOnStage<'computed'>;
export type AnyLayouted = AnyOnStage<'layouted'>;
export type Any = Aux<any, any, any, any, any, any>;
// ...
export type Fqn<A> = A extends Any ? scalar.Fqn<ElementId<A>> : never;
/**
 * Element FQN from Aux as a literal union
 */
export type ElementId<A> = A extends Any ? Coalesce<A['ElementId']> : never;
// ...
export type ElementKind<A> = A extends Any ? Coalesce<A['ElementKind']> : never;
/**
 * DeploymentKind from Aux as a literal union
 */
export type DeploymentKind<A> = A extends Any ? Coalesce<A['DeploymentKind']> : never;
/**
 * RelationKind from Aux as a literal union
 */
export type RelationKind<A> = A extends Any ? Coalesce<A['RelationKind']> : never;
/**
 * Tags from Aux as a literal union
 */
export type Tag<A> = A extends Any ? Coalesce<A['Tag']> : never;
/**
 * Array of tags from Aux
 */
export type Tags<A extends Any> = readonly Tag<A>[];
/**
 * Metadata key from Aux
 */
export type MetadataKey<A> = A extends Any ? Coalesce<A['MetadataKey']> : never;
/**
 * Metadata object from Aux
 */
export type Metadata<A extends Any> = IsNever<A['MetadataKey']> extends true ? never : IsLiteral<A['MetadataKey']> extends true ? {
    [key in A['MetadataKey']]?: string | string[];
} : Record<string, string | string[]>;
// ...
export type OrString = string & Record<never, never>;
/**
 * Allows any string value, but still auto-completes to the possible values in IDE
 */
export type LooseLiteral<V extends string> = Coalesce<V> | OrString;
export type LooseElementId<A> = A extends Any ? LooseLiteral<A['ElementId']> : string;
export type LooseDeploymentId<A> = A extends Any ? LooseLiteral<A['DeploymentId']> : string;
export type LooseViewId<A> = A extends Any ? LooseLiteral<A['ViewId']> : string;
export type LooseTag<A> = A extends Any ? LooseLiteral<A['Tag']> : string;